requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pydantic-settings
streamlit

# 랭체인 메인 (체인 생성 함수 포함)
langchain

# 핵심 코어 및 커뮤니티 (로더 등)
langchain-core
langchain-community
langchain-classic

# 모델 및 DB 연동
langchain-openai
langchain-chroma

# 텍스트 분할 도구
langchain-text-splitters

# 랭그래프 (LangChain → LangGraph 전환용)
langgraph==1.0.8
langgraph-prebuilt==1.0.7

# 기타 필수 의존성
chromadb
openai
tiktoken
python-dotenv
rapidfuzz

# 키워드(BM25) 검색용
rank-bm25

# RAG 평가(RAGAS)
ragas
datasets

# 랭그래프 아스키 코드로 그릴때 필요
grandalf

# Cross-Encoder 기반 리랭킹(문서-질문 쌍 점수화)
sentence-transformers

# PDF 파일 읽을떄 필요함. 랭체인 내부에서 쓰이는듯.
pypdf

# 뉴스 검색
tavily-python

# 베이지안 최적화 (하이퍼파라미터 튜닝)
optuna

# 테스트
pytest

# DSPy (프롬프트 자동 최적화 프레임워크)
dspy

# Optuna 결과 시각화 대시보드 (optuna-dashboard sqlite:///optuna_results.db)
optuna-dashboard